removing daq_scan_ui to simplify coding
DQ_Scan is a CustomExt so it is a UI no need to add another layer
We already have DAQ_Scan_Acquisition to perform the scan mechanism
